PixelMinds/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
pydantic==2.7.0
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
chromadb==0.5.0
sentence-transformers==3.0.0
alembic==1.13.0
